refactor(iam): split sts_service.rs by surface (3095 -> 1840 mod + 4 submodules)
Audit 2026-05-19 P9 continuation. sts_service.rs split:

- assume.rs       870  (AssumeRole / AssumeRoot path)
- federation.rs   200  (AssumeRoleWithSAML / AssumeRoleWithWebIdentity)
- session.rs      139  (GetSessionToken / DecodeAuthorizationMessage)
- caller.rs        85  (GetCallerIdentity / GetAccessKeyInfo)

mod.rs (1840) retains the AwsService trait impl, dispatch, the
shared helpers, and the (large) test corpus.

299 IAM unit tests pass.
